Fix package name warning
