+++ src/main/java/adam/parser/ListOnCommand.java
@@ -1,5 +1,6 @@
-package adam.input_handler;
+package adam.parser;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 
 import adam.core.TaskList;
@@ -7,9 +8,17 @@ import adam.core.Ui;
 import adam.exceptions.AdamException;
 
 /**
- * Represents a command to list all tasks in the task list.
+ * Represents a command to list all tasks on a specific date.
  */
-public class ListCommand extends Command {
+public class ListOnCommand extends Command {
+    /** The date to list tasks on */
+    private LocalDate date;
+
+    ListOnCommand(String input) throws AdamException {
+        super();
+        this.date = Parser.parseInputDate(input.split(" ")[1]);
+    }
+
     /**
      * Checks if the input matches the command.
      *
@@ -17,11 +26,12 @@ public class ListCommand extends Command {
      * @return True if the input matches the command, false otherwise.
      */
     public static boolean isMatch(String input) {
-        return input.equals("list");
+        String[] inputParts = input.split(" ");
+        return inputParts[0].equals("listOn") && inputParts.length == 2;
     }
 
     /**
-     * Lists all tasks in the task list.
+     * Lists all tasks on the specified date.
      *
      * @param manager The task list to add the task to.
      * @param ui The user interface to output to.
@@ -29,7 +39,7 @@ public class ListCommand extends Command {
      */
     @Override
     public void execute(TaskList manager, Ui ui) throws AdamException {
-        ArrayList<String> outputs = manager.listAll();
+        ArrayList<String> outputs = manager.listAllOnDate(this.date);
         for (String output : outputs) {
             ui.outputText(output);
         }
